+++ src/components/form/signUpForm/signUpForm.module.scss
@@ -1,0 +1,11 @@
+.signUp {
+  margin-top: 2rem;
+
+  input {
+    padding: .2rem .4rem !important;
+  }
+}
+
+.submitBtn {
+  width: 100%;
+}
