+++ src/components/login/altSignIn/altSignIn.module.scss
@@ -1,27 +1,33 @@
-.alternativeSignIn {
+.altSignIn {
   width: 100%;
   display: flex;
   flex-direction: column;
-
-  p {
-    margin: 0 1rem;
-  }
-}
-
-.orSeparator {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 1rem;
-}
-
-.orSideLine {
-  width: 40%;
-  height: 1px;
-  background-color: var(--color-grey);
 }
 
 .buttonsGroup {
   display: flex;
   flex-direction: column;
+
+  button {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    opacity: .9;
+    font-size: var(--login-page-sign-in-alt-buttons-font-size);
+    color: var(--login-page-sign-in-alt-buttons-color);
+
+    &:not(:last-child) {
+      margin-bottom: var(--login-page-sign-in-alt-buttons-mar-bot);
+    }
+
+    &:hover {
+      opacity: 1;
+    }
+
+    svg {
+      width: var(--login-page-sign-in-alt-icon-size);
+      height: var(--login-page-sign-in-alt-icon-size);
+      margin-right: var(--login-page-sign-in-alt-icon-mar-right);
+    }
+  }
 }
